refactorization with publisher segregation
