+++ pymod/run.py
@@ -5,6 +5,7 @@ import sys
 import time
 
 from argo_nagios_ams_publisher.threads import Purger
+from argo_nagios_ams_publisher.publish import Publish
 from collections import deque
 from messaging.error import MessageError
 from messaging.message import Message
@@ -101,26 +102,3 @@ class Run(object):
             self.log.error(e)
 
 
-class Publish(Run):
-    def __init__(self, *args, **kwargs):
-        for d in kwargs.iterkeys():
-            code = "self.{0} = kwargs['{0}']".format(d)
-            exec code
-        self.init_confopts(kwargs['conf'])
-
-    def write(self, num=0):
-        published = set()
-        try:
-            for i in range(self.pubnumloop):
-                with open('/root/msgs_file', 'a') as fp:
-                    fp.writelines(['{0}\n'.format(str(self.inmemq[e][1]))
-                                   for e in range(self.msgbulk)])
-                published.update([self.inmemq[e][0] for e in range(self.msgbulk)])
-
-                self.inmemq.rotate(-self.msgbulk)
-
-            return True, published
-
-        except Exception as e:
-            self.log.error(e)
-            return False, published
